requirements.txt
Pillow>=9.0.0
argparse
python-magic>=0.4.24
psutil>=5.9.0
colorama>=0.4.6
pytsk3>=20220519
rich>=13.0.0
questionary>=1.10.0
prompt-toolkit>=3.0.0
click>=8.0.0
pyyaml>=6.0
tqdm>=4.65.0
cryptography>=3.4.0
python-dotenv>=0.19.0
